hub: move csmock_parser.py module from common to hub
Resolves: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/268
