+++ osh/hub/service/loading.py
@@ -9,7 +9,7 @@ import logging
 
 from kobo.hub.models import Task
 
-from osh.common.csmock_parser import CsmockAPI
+from osh.hub.service.csmock_parser import CsmockAPI
 from osh.hub.service.path import TaskResultPaths
 
 logger = logging.getLogger(__name__)
