+++ osh/hub/waiving/results_loader.py
@@ -8,8 +8,8 @@ from django.core.exceptions import ObjectDoesNotExist
 from kobo.hub.models import Task
 
 from osh.common.constants import DEFAULT_CHECKER_GROUP
-from osh.common.csmock_parser import CsmockAPI, ResultsExtractor
 from osh.hub.scan.models import AnalyzerVersion, AppSettings
+from osh.hub.service.csmock_parser import CsmockAPI, ResultsExtractor
 from osh.hub.service.path import TaskResultPaths
 from osh.hub.service.processing import (TaskDiffer, task_has_results,
                                         task_is_diffed)
